refactor(video): Refactor VideoLibraryScreen for Enhanced PiP Mode Handling (#81) (fixes #76)
